Store decimals

+++ database/migrations/2020_01_27_110327_create_payments_table.php
@@ -16,7 +16,7 @@ class CreatePaymentsTable extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('object_payment', 160);
-            $table->float('amount');
+            $table->decimal('amount', 15, 2);
             $table->unsignedBigInteger('payment_type_id');
             $table->unsignedBigInteger('payment_method_id');
             $table->unsignedBigInteger('state_id');
